Added database, and overall server management system
Added the Manager class, that is used to create servers, create clients, and manage queues. Added 3 tables to the database, for users, game servers, and instances of servers. Moved database connection config to a seperate file, config.cfg, which is hidden for security purposes. Removed client connection functionality from the Server class, as it is now part of the Manager. Server class now pulls config options from the database.