Added documentation
Added mkdocs, the system that will be used to generate documentation pages from markdown. Added the material theme (with some alterations) for mkdocs. Added an intro page and an installation guide. Added the html file (hard symbolic link) for the API docs page, as well as all of the fonts that it requires. Made the home page slightly wider to match the documentation. Made the navbar a seperate template that is included by the home page. Flask debug mode is active when testing locally. Added all of the necessary Flask code for the new documentation.